Restructure portable to target-json
