+++ src/ilang++.cc
@@ -5,8 +5,8 @@
 
 #include <ilang/config.h>
 #include <ilang/ila/instr_lvl_abs.h>
-#include <ilang/portable/interface.h>
 #include <ilang/synth-interface/synth_engine_interface.h>
+#include <ilang/target-json/interface.h>
 #include <ilang/util/log.h>
 #include <ilang/verification/abs_knob.h>
 #include <ilang/verification/unroller.h>
